kills ptpe
Following the idea that ensugaring matters don't belong in Host,
I'm removing the last (huge!) piece of ensugaring - ptpe - which converted
scalac's types to palladium's trees.

At the moment, I'm just removing this functionality altogether and putting up
a stub called Type.syntaxize that in the future will provide analogous logic.
Later on, when implementing Host.attrs, we'll have to implement it.