moves supercall ensugaring to host
This was one of the few ensugarings when I hesitated wrt their placement.

On the one hand, representing `B(2)` in `class C extends B(2)` as an Apply
is syntactically natural, even if a bit awkward.

On the other hand, applications of types to terms are particularly incompatible
with scalac's infrastructure (e.g. app.symbol is a type symbol, not a superctor symbol),
and that bit me when I implemented name/default desugaring for supercalls.
Therefore, I'm reverting the original decision to have supercall processed
in ensugar and moving it to host.