foo(bar) => foo.apply(bar) is now processed in the converter, not in the ensugarer
Otherwise, we end up with really awkward Apply trees that don't have symbols
(because the tree representing `foo` doesn't have to be a SymTree).

This is probably going to bite us in the future, like it's already happened
in interpolation desugaring (now we have to account for both
`StringContext.apply(...).q(...)` and `StringContext.apply(...).q.apply(...)`),
but I only have time for worrying about this later.