travis will no longer publish built artifacts
Reason: this way we can get binaries for unpredictable JVMs published
to Sonatype, and that becomes a problem, because scalahost.jar embeds
scalameta.jar and we need those classes to be identical. This can definitely
be fixed with a bit of build hackery (e.g. we could prohibit publishing
on all JVMs except for 1.6), but I just don't have time for that now.

For now, it's not a problem, because I run publishing manually at
appropriate moments of time. If our workflow becomes more intense,
I'll revisit this.