temporarily removes monadic error handling UI
This is something that Denys and I discussed quite a lot last spring.
We haven't arrived at an agreement wrt how fallible APIs should be expressed
(via exceptions or via monads), so we implemented a compromise solution,
which abstracted away error handling behind an implicit strategy, which
the user could then specify.

Along with the benefit of flexibility that would enable us to do practical
experiments on error handling strategies, this solution also brought
significant notational overhead. For example, a method that transformed
a seq of types to a seq of classes they point to, required 10+ loc [1]
as well as a 10+ loc helper [2] to get job done instead of straightforward <5 loc.

While I do think that monadic error handling is an interesting idea,
which is definitely worth experimenting with, in its current incarnation
it's going to (telling this from my own experience!!) significantly slow down
the implementation of semantic API (even though our users could just import
`errors.throwExceptions` if they didn't want to pay for extra encapsulation
and compositionality, we were forced to use monads to be compatible with
both error handling styles).

Therefore I'm temporarily rolling back our error handling UI to use exceptions.
I'm not entirely happy with this decision (though the relief of keeping code
simple is real), so looking forward to the day when we can experiment with this
again and maybe find better syntactic ways to tap into the power of monads.

[1] https://github.com/scalameta/scalameta/blob/8bf99afa43e2ab1181573c50a12cd321c884019f/reflection/semantic/package.scala#L51
[2] https://github.com/scalameta/scalameta/blob/8bf99afa43e2ab1181573c50a12cd321c884019f/foundation/errors/ErrorHandler.scala#L16