introduces the notion of dialects
Dialects are meant to encapsulate minor differences between language variants
ensuring that we can use the same APIs and infrastructure to work with various
Scala-like languages.

Immediate motivation for introduction of dialects is the desire to provide
correct prettyprinting for Dotty vararg patterns. The thing is that Dotty
has changed the syntax of binds to sequence wildcards from `List(xs @ _*)`
to `List(xs: _*)`, and now we need AST prettyprinting to be able to work well
both for Scala and Dotty.

Therefore, I've added the `scala.meta.Dialect` trait (that only has a single
method for the moment) and changed all prettyprinting, syntactic and semantic
facilities to take an implicit parameter of type `Dialect`. Handling of the
motivating difference between Scala 2.11 and Dotty will arrive in a subsequent commit,
and here we focus solely on the mechanism that allows for that.

Allright, so essentially every scala.meta API (except for tree construction and
deconstruction) takes an implicit dialect. Now where do we get those implicit
values from? These values be obtained in one of the following ways:
1) By having an implicit of type `scala.meta.xxx.Context` in scope
(both `semantic.Context` and `macro.Context`, as well as the upcoming
`syntactic.Context`, provide `Dialect`).
2) By importing one of the `scala.meta.dialect.XXX` implicit objects.

In a nutshell, this means that in order to use scala.meta APIs, you need to
either:
1) Have a syntactic/semantic Context (and that context will take care
of all dialect-specific stuff, because contexts are provided by
platform-specific hosts, which is aware of their dialects).
2) Manually choose the dialect (because we can no longer say
"I want to parse this file"; we need to be more specific as in
"I want to parse this file as Scala 2.11 code").