Member.isPkg/isPkgObject => isPackage/isPackageObject
In my opinion, Package is vastly superior to Pkg (after all, we don't say
Cls, Obj and Trt, but Class, Object and Trait).

Unfortunately, we couldn't have scala.meta.Package, because that would
conflict with scala.meta.package. Now that we don't have Pkg in the public API,
we don't have to be consistent and say Member.isPkg.