casts to internal! yay!!
If we want underlying tree classes to be potentially host-defined,
we'll have to find a way around this.

The original plan was to use quasiquotes that desugar into host-defined
tree constructors, but we currently can't use quasiquotes when all that
we have is semantic.Context (because its dialect is of type Dialect,
which is not enough to parse code at compile-time). We'll get around this
somehow, I'm sure, but that's a story for another day.