Template.stats is now Option[Seq[Stat]]
The type is incredibly ugly, but we do need a trivia-less way to distinguish
`new X` and `new X{}`. It's easy to imagine something like splitting
Term.New into Term.New and Term.NewAnonymous, but then the distinction
between those wouldn't be expressible via quasiquotes.

Therefore, for now, I'm just going with the simplest of the two evils.
Later we will want to revisit this decision