haha! semantic quasiquotes don't need a semantic context
At least for now, i.e. for the case when we are only going to resolve global symbols.
Global symbols are supposed to be stable across all hosts, so it doesn't matter
who creates them - a host or code emitted by a reifying macro!

We might also be fine without hosts even for local symbols, because for those
a reifying macro can just take a look into their persistent IDs.
Things get a little messy, when a quasiquote isn't precompiled, but is instead interpreted,
but I think that everything's going to bootstrap just fine (because an interpreter
is run by a host, so an interpreter can coordinate the symbols that it sets to quasiquoted names
with the host).

This reverts commit 0db9820db2549a6123e2121e13fb3ff6262f1ab1.