templates are no longer scopes
It's nice and all that from the convenience standpoint, but unfortunately
we can have only one of two things:
1) class scope includes its template as a nested scope, but doesn't
include the members of the template
2) class scope includes its template members, but template is not a scope
(because having members belong to multiple scopes is sorta weird).

I'd much rather be able to write `class.members` and immediately get
to its members (#1) than be able to write `template.members`, but have
to just an extra hoop to get to class members (#2). That's why I'm unmaking
template a scope.