Term.This and Term.Super are now names
We need to remember their denotations, and one way of doing that is
to turn them into names. Then we wouldn't need to change our mantra that
"everything binding related is expressed in and by names".

And, you know what, `this` and `super` definitely look like special names,
so it's not a stretch to say that they are names. They are definitely
not term names or type names, just names.

Maybe this decision will backfire - who knows - but let's wait and see
how this pans out.