Member.ref => Member.name
Oh my god! My dream from last spring has finally come true!!

Since all references to members are now names, we can rename Member.ref to Member.name.
This goes along the lines of the idea that everything binding-related
is represented with names.

The only fly in the ointment so far is the existence of anonymous term
and type parameters, which will crash Member.name (the very same way
that they crashed Member.ref previously). To be honest, I'm thinking
of creating something like Name.Anonymous to avoid the crash.