introduces Name.Anonymous
This may sound nuts, but absense of a name, as in e.g. `def foo[T[_]] = ???`
is not good enough reason to avoid creating a name.

Much like in scalac, we need to create symbols for anonymous identifiers,
we need to create names for anonymous identifiers as well. It is true
that we don't need to compare those names, and neither we need to reference
those names (because, after all, they are anonymous). However, we also
need to track host-to-native correspondence for the definitions that these
names stand for, and the only mechanism that we have for that is bound
to names.

Why not create a new mechanism then? Well, because that will dilute
the very important principle that I've been musing on in the few previous
commits. Everything binding related should be handled in and by names.

I realize that all this fun comes at a cost of losing some safety.
In order to make Member.ref (or, in new terms, Member.name) useful,
Name.Anonymous has to be both Term.Name and Type.Name (otherwise, everyone
who wants to unquote a reference to a definition would have to first check
whether that reference represents an anonymous name).

Good news, though, is that: a) no compile-time safety is lost
(i.e. all programs that compile now were going to compile before this commit),
b) safety by construction isn't violated either (i.e. it is still impossible
to construct an invalid tree, e.g. a def with an anonymous name).
The only degradation in experience comes from the fact that mistakes
that would be caught by a crash in Member.name are going to survive
until this name is unquoted somewhere, which definitely increases confusion.