Context.owner is gone, Context.members got a more specific type signature
It turns out that Context.owner/Context.members was a false symmetry.
We don't need owner at all, because we can compute it from tree's parents
(even members returned from a host have a parent set, and those who don't
have it set for their Member.source).

Context.members is only a host-based operation for types - for all other trees
we can compute the list of members syntactically (or, in case of templates,
by delegating to members on Type).