a sad discovery
Not all post-typer types are representable with syntax.
Just found a counter-example in Typers.scala, which was hidden under a bug
in our g.Type => p.Type converter.

I believe we can have an adequate emulation of the counter-example
as described in the comment, but we'll have to see whether the emulation
can be well-received by scalac.