introduces meta.Pat.Type
Now I remember why we didn't introduce Pat.Var a year ago.

Because it would require a symmetric change to express type variable patterns,
and that would make it impossible to use meta.Type types for type patterns.
That was deemed too inconvenient and rejected.

Now I think that we should've followed that through, because pattern types
seem to form a distinct syntactic group, with their own peculiarities.
An argument that types and pattern types look the same syntactically
is compelling and seductive, but Term.Apply and Pat.Extract also look
the same, and we didn't unify them, so I think that there is more to
tree unification that just syntactic likeness.

Even though this change creates an inconvenience (it's no longer possible
to use Type and Pat.Type interchangeably), that inconvenience is arguably minor,
because one could always imagine having a bidirectional conversion, with
the Type => Pat.Type direction being always successful (+ a Liftable)
and Pat.Type => Type being potentially fallible.