Term.Name and Type.Name are no longer members
This was a necessary measure when naked names could be binders in patterns,
but now when we have Pat.Var classes, we don't have to do this anymore.

I think this is a positive change, because it disambiguates semantic API:

1) In quite a lot of places, when working with members, I had to write
things like `case name: Term.Name if name.isBinder`, and that was noticeably
inconvenient and definitely easy to forget.

2) Also, Term.Name and Type.Name are prominent members of our API, so
having them as kinda members invited code that compiled, but made no sense.
E.g. I spent the best part of the 22nd hour yesterday trying to understand
why `t"List".members` works, but `q"List".members` produces an empty list.
Now it's obvious: q"List" was returning Term.Name, which counted as a Member
and therefore had an extension method `members` on it.