changes Member.tpe to be Type, not Type.Arg
This reflects the fact that `=> T` and `T*` are not real types,
and that even references to parameters that have those types are typed differently.

Incidentally, this makes Member.tpe to be roughly 10x more useful,
because one doesn't need to pattern match Type.Args away and can utilize
that method's result right away.