introduces Lit.Byte and Lit.Short
We can't cop out with ??? anymore, because we need to represent types of byte
and short constants in order to successfully read scala-library.jar.
Therefore, I decided to take a stab at adding these two new flavors of literals.

There are two problems that need to be solved here, and below I'll explain
how the are addressed in the pull request:
1) What syntax are we going to use to construct and deconstruct these lits?
2) How do we prettyprint these lits?

Both questions actually have reasonable answers:
1) No surface syntax, but we can use lifting/unlifting.
2) We can delegate code representation of those literals to a standalone library.
   The toughest thing here was to find a notation that would work for
   both terms, types and patterns, but I think I've succeeded.