parameters are no longer Member.XXX, but just Member
As long as Term.Param was Member.Term, we couldn't have Member.Term.name
return Term.Name, but had to settle on Name with Term.Ref (because params
represent self types, and self type's name is a This). Therefore, I demoted
Term.Param to Member (and Type.Param followed it to be consistent).

This allowed to address another thing that I didn't like - the fact that
Name.Anonymous was Term.Name and Type.Name. Now Name.Anonymous is just Name,
which is perfectly fine, because parameters are just Members.