principled handling of fringe names
Bites the bullet and introduces 6 new subclasses of Name into the public API
and 5 new subclasses of Name into the internal API to handle:
* Parameter names
* Imported names
* Access boundary names

An alternative is to have all those represented with just Name in the public API
and then to have runtime validation when these names are unquoted in.
That way we simplify the public API quite a bit (by removing more than
1/6th of all the traits there), but sacrifice static safety and understandability
in corner cases (because people won't know what to unquote in there).

After thinking through it, I'm willing to give the former solution a try,
because I do like static safety here. We can always cut the "complexity"
if it ends up being too much.