standard Ctor.Ref <=> Type conversion methods
As long as we're still on the fence wrt the exact representation of ctors,
we need some standard way to convert between types and ctor refs.

Ctor.Ref => Type is already taken care of by the means of .tpe,
whereas for Type => Ctor.Ref we need to do some work.