compatibility with macro paradise
Scalahost hijacks the analyzer in order to be able to attach additional
metadata to trees during typechecking, and that messes up analyzer and macro plugins
that by that time have already been registered in the old analyzer.

This is a fundamentally unsolvable problem, so we'll need to migrate away
from analyzer hijacking, but for now, while we still can't do that,
I'm putting a workaround in place that makes scalahost compatible with
macro paradise, the only widely used analyzer/macro plugin (to the best of my knowledge).