improves enterPluginStats workaround's performance
It turns out that getStackTrace is much more expensive than I imagined.
After the stacktrace-based workaround was merged, `sbt compile` on scalaz,
which doesn't need the workaround in the first place, got 1.5x slower.

I've hacked up a quick patch that brings the slowdown down to 1.1x
on projects that don't use macro paradise. Let's hope that my fix will
get merged in 2.11.6, which is to be released soon, and we won't need
the workaround anymore.