prepares the atPos/autoPos infrastructure
Looks like we won't be able to automatically derive positions for parsed trees,
because it sometimes happens that a tree is parsed as something else
and then that something else is reassembled into the final result.
Therefore, we'll need atPos anyway.

However, I think that the majority of trees will be able to be positioned
automatically, so I'm also adding autoPos, which simply remembers the
positions at the beginning of the code block and ar the end and then uses
them to attribute the execution result.