Revert "prohibits atPos from overwriting existing positions"
This reverts commit f05a9e37b2770799f29d6ed7862794b446562329.

So the question is: should parentheses be part of ApplyInfix in, say, `(2 + 3)`?
The commit that is being reverted right now says "no", and for that purpose
it changes `atPos` to not overwrite existing positions, keeping ApplyInfix at `2 + 3`.

However, if we go for it, what do we do with stuff like `(2 + 3).something`?
How do we let Term.Select know that it needs to expand itself beyond its qualifier?
In the current architecture, that'll be a very hard thing to do, so let's go
for including parentheses into that ApplyInfix.