@ast root and branches are no longer sealed
I'm about to introduce impl.Unquote, which has to be a subtype of all tree nodes.
However unpleasant the consequences are, we have to deal with them, because
we currently don't know of an alternative technology to implement quasiquotes
in a principled manner.

What we discussed with Denys is that we might be able to get back the
guarantees offered by sealedness if we find a way to customize exhaustivity
checks in the pattern matcher from a compiler plugin. People are going to
use scalahost anyway, so it could also do additional domain-specific
exhaustivity checking.

This commit also comes with some refactoring of the adt/ast infrastructure.
In particular, all checks now return q"()" instead of q"", which has uncovered
a terrible bug. If a trait def has the INTERFACE flag set, then essentially
any statement in its body (e.g. q"()") is going to crash the compiler.
Therefore, I'm manually resetting the INTERFACE flag in @root and @branch.
This, in turn, led to problems with `@branch trait Defn`, because it makes
scalac emit `Defn$class`, which comes in conflict with `Defn$Class`.
To address that, I had to temporarily disable all checks for Defn.