adt/ast annotations no longer reset INTERFACE flag
As explained in the parent's commit message, we had to reset INTERFACE,
because our codegen emits statements into templates of the annottees,
and GenICode really doesn't like it when a template of an INTERFACE trait
has statements in it.

Now it's time for a principled solution. All statements are now moved
to companions of the annottees, and everyone's happy.