turns all @ast classes into traits
I'm about to introduce scala.meta.internal.ast.Unquote, which has to be
a subtype of all tree nodes. For that, we need to make it possible to
subclass @ast classes.

The solution that I'll be pursuing in this and subsequent commits
is splitting `@ast class Foo` into `trait Foo; object Foo { class Impl extends Foo }`.
This is a very heavy-handed way of organizing the @ast codegen, but it looks
like we don't have much choice.

A pleasant side-effect of this change is the ability to proxy trees,
so that, for example, we can lazily load bodies of vals/vars/defs.
Nothing in that direction is implemented just yet. We will cross that bridge
when we get there - now is not the time.