Term.This/Term.Super: Option[String] => Name.Qualifier
This takes care of the last place in the tree API, where we represented
names with strings, and makes our tree API maximally friendly to Unquote.

It's a bit of stretch to use Name.Qualifier here, because along with
anonymous and indeterminate names (which fit), Name.Qualifier also covers
Term.This (which doesn't fit), but I didn't feel like introducing a new
public trait just to have this extra bit of static safety. Maybe in the future
we'll have to change this decision.

Having this change in place also means that This and Super don't have to be
names anymore. This, in turn, means that Name.Qualifier can no longer be
a Name and becomes a Ref. This also means that Member.name can no longer
be a Name, but that was too much for me, so now it's still a Name, but it
does not return This anymore. All in all, this worked pretty well.
I'm not sure whether I like the fact that `this` and `super` are no longer names,
but I can try to get accustomed to that.