Input is no longer sealed
The prototype of Input, previously called Origin, was sealed, because it
doesn't make sense for, effectively, a position to have an unlimited
number of implementations.

However, now that we've split Origin into Input and Origin proper,
we can unseal the input-abstracting part, allowing everyone to provide
input to the parser, as long as they can yield an array of characters.
Later on, we may make this collection of characters lazy.