prepares the parser to accept precomputed sequences of tokens
Parser and AbstractParser are now merged into an entity that takes
an input along with an implicit dialect. CrazyTokenIterator just takes
input.tokens and does its thing. All variations in tokenizing the input
should now be handled outside the parser.

There's now Input.Slice and Input.Tokens, which express the use cases
that are necessary to parse quasiquotes, namely: 1) taking part of other input
and then tokenizing it, 2) taking a precomputed sequence of tokens and
feeding it to the parser. Note how all that is encapsulated in Input,
and the contract of both parser and tokenizer remains intact.

Even though I'm talking about sequences of tokens, the only kind of sequence
that is currently supported is Vector[Token], just like the only representation
of Input.content that is currently supported is Array[Char]. Maybe in the future
it'll make sense to make either or both of them lazy.