introduces Token.prev and Token.next
I've long considered adding a way for tokens to access their successors,
but there never was a pressing need for that, so I've never went for it,
aiming to have a solution that's as simple as possible.

Today, the time has come. In order to figure out whether an Ellipsis
or an Unquote introduce a definition (e.g. is[DefIntro]), we need to know
whether they are followed by a DefIntro token. In order to implement this
logic in Token (and this kind of stuff has been neatly encapsulated in Token
until now), we need to have Ellipsis and Unquote know their successors.

A simple solution here would be to add an index: Int field and then
use parent input's tokens collection to navigate the prev/next links.

Unfortunately, it's not that easy. First, now that we have index,
we need to clone all tokens that are encapsulate in Input.Tokens, and
in order for those tokens not to lose their positions, we need to add
Token.prototype. Second, just index is not enough, because we also need
a dialect in order to access parent's tokens.

All in all, it worked fine, but I'm very annoyed by the fact that tokens
have grown from lightweight constructs into monsters that carry around
a bunch of stuff. We'll have to optimize that in the future.