conditional is[T] for Token.Case
Makes use of the prev/next infrastructure to make Token.Case conditionally
conform to TemplateIntro and CaseDefEnd. Also introduces CaseIntro to further
simplify the parser.

It wasn't a trivial change though, because I had to change another piece
of infrastructure. With conditional is[T], `_: T` and `is[T]` are no longer
interchangeable. More precisely, we can no longer do `_: T` for branch traits.
We'll have to do something about this, probably by migrating to full-fledged
typeclasses instead of this half-assed approach.