conditional is[T] for Ellipsis and Wildcard
To tell the truth, I don't like this, because overloading is[T] for these tokens
conflates uses of is[T] for dispatch (that's okay for us) and for generalizing
over tokens (that's not okay). I think it would be better if we just
introduced additional conditions in places where we have relevant is[T] checks.