introduces Mod.Ffi
I can't say I like this idea very much, but we need an immediate solution,
so that scalahost can represent abstract methods that originate from bytecode.
In the case of concrete methods, we could replace their bodies with stubs that carry
all the necessary info, but with abstract methods there's not much that we can do.
We will need to revise this decision later on when the dust settles.