another hole in our representation of constructors
Unfortunately, as things stood, we couldn't write Type.ctor.params,
and that was a disaster.

We can't even change Scope.ctor to return Term.Ref instead of Term,
because that's not going to work with things like `new Foo[T]`
which are represented as a Term.ApplyType over a Ctor.Ref.

Therefore, I'm downgrading Scope.ctor back to Member.Term, leaving the joy
of calling `tpe.ctorRef(ctor)` to the users.