one step closer to toApproximateMtree
This is necessary to deal with conversion of macro application arguments,
some of which can be inferred implicits, created with mkAttributedRef
that creates desugared code out of the blue.

Another use for this patch is the upcoming implementation of Term.desugar,
which I'll have to handle next week for the interpreter to work.