updates ReflectionSuite
Note the huge increase in the number of branches and leafs.
That's because for every Foo, we not generate Foo.Quasi, Foo.Unquote and Foo.Ellipsis.
Unsurprisingly, this also bumps the size of the generated binary from 4.2 Mb to 7.9 Mb.

This is definitely too much, so we'll have to explore possible optimizations.
E.g. we could merge Foo.Ellipsis into Foo.Unquote, having a single AST node
called Foo.Unquote with rank >= 0.