finally defeats the "continuing with the stub" problem
Apparently there was a problem with enumerating members of the scala package
within the context of the batch compiler (e.g. when expanding macros).

When doing something like q"scala".members, we would get the following
unsquelcheable warning: "warning: Class scala.languageFeature$experimental not found - continuing with a stub".
I believe that this is specific to GlobalContext, because back then I couldn't repro
it under StandaloneContext (it's come up before, but I just didn't know how to fix it).

Apparently, this warning is emitted when ClassfileParser tries to parse
a *.class file for a fake ClassSymbol created by the batch compiler when
only a *$.class file exists. Why would it parse the non-existent class file?
Well, apparently, this somehow happens when we do Symbol.exists.

After I moved the existence check after a synthetic name check, everything became normal.
Would be nice to debug wtf scalac is thinking, but that's the story for another day.