fixed a fundamental bug in function calls
We can't use function environment as an evaluation result of a function call.
We must keep on using our own environment, with our own lexical scope,
but enriched with side effects from the function environment.

Since we don't model side effects yet, we simple replace functionEnv with callEnv,
and then enjoy the fix :)