BF: covering more corner cases
The converted tree sometimes contain more information than the parsed one in
its own structure:
for instance, Updates and Tuples are already converted as such. Sometimes it
happens that name prefixes are added, transforming Names into Select. Some of
those information are interesting at the semantic level, so we transfer them
to the parsed tree. We do that only when required, as we want the form of the
parsed tree to prevail in case of non classified differences.

I am more and more concern regarding the fact that this might not be a viable
solution for the longterm. As you (Eugene)  mentionned, it might be better to
parse the tree and add semantic information form scalac directly on top of it
rather than converting it. The current merging of trees relies on many layers
layers:
first, the conversion by toMTree must not leave any ambiguity, and then some
ambiguities might be added by the fact that the two tree differ during the
merging. It does the trick for tools like Obey while running on complex
projects (Tested on various Play apps), but it might return warnings in some
corner cases were not covered. This would have to be tested more thoroughly.