refactors origins
Origin.None and Origin.Transformed got merged into Origin.Synthetic
that possesses the capability (TBI) to infer its dialect and tokens.

I think that Origin.Transformed will be useful in the future, but so far
we're lacking experience with that kind of stuff, so I'm just removing it,
because it was unused since its very inception.