refactors inputs and positions
1) Inputs are now sealed and split into Input.Real (the ones that have
content, these are still unsealed) and Input.Virtual (the ones that only have
tokens). This exposed a bug in scannerTokenCache - types for the win!

2) XXX.start/end are now Point instead of Int, and two points combine into Position.
Much like inputs, points can be real and virtual (there's no use of having
a position into a virtual input, because it can be combined from multiple inputs!).

In the ideal world, these would be two separate commits,
but I really want to go for a run earlier than yesterday.