adds Token.dialect
I'm about to remove Origin and replace it with Tokens, but for that we'll need
to remember dialects somewhere, and that somewhere can be Token.

This temporarily increases memory footprint of Token, but later on,
when applying the TODO optimization in Tokens.scala, we'll be able to
optimize both Input & Dialect pointers away, as easily as we would optimize
the Input pointer.