refactors Exceptions to take positions instead of inconsistent data bunches
This was harder than I initially expected, because this change became
a battle test of the position API that I prototyped in the previous commit.
What can I say... The prototype didn't work well :)

The key takeaway is that it's not enough to say that a tree that's parsed
from a mishmash of tokens (i.e. not from Content, but from Tokens)
doesn't have a position (i.e. has Position.None). When we need to report errors
for such trees, we need a point, and even if we don't have a range position
here, we still can expose a useful point.

A pleasant side-effect here is that by exposing Position.Assorted this we get rid of
null objects for inputs, positions and points, which is simply wonderful!

Also, I can definitely say that having a position whose input is different
from its token inputs is confusing. I'll be keeping an eye on this.