experimenting with a new approach to structuring the converter
I tried to write a converter that handles a simple example provided to me
by @kolowheel (a package with a couple classes and a simple method),
and it ended up being an entangled mess that actively resisted against
me trying to factor out some of the parts into separate functions.

Upon some reflection, I'm thinking that the reason is the particular design
of scala.reflect trees. 1) scala.reflect trees don't have parent links,
so one always has to pass some sort of context around in order to be able
to compute or remember info about parents. 2) scala.reflect trees are
irregular (mods are not trees, names are not trees), so one always has
to process these special cases by creating special cases of their own.

The idea is now to have an intermediate representation that is half-scala.reflect
and half-handrolled in order to offset the two aforementioned issues.
Hopefully this will allow to modularize the new version of the converter.