changes unfortunate names in Typing and Expansion
Typing.Unknown was meant to denote a not-yet-filled-in typing, not a typing
that's unknown due to some reason. A typing might be known, but not yet
filled in.

Typing.Known only made sense when there was Typing.Unknown. Once the latter was gone,
I had to change the former as well.

Expansion.Identity was meant to denot a not-yet-filled-in expansion,
not an expansion that's known to be an identity.