introduces type lambdas
When working on the new incarnation of the scala.reflect => scala.meta converter,
I had to solve a problem that's been glossed over in the previous implementation:
the problem of assigning types to terms.

Previously, we had a great cop-out approach that consisted in attaching
native reflection artifacts to scala.meta trees. For instance, if we converted
a scala.reflect term, then we would created a scala.meta term with the original
term attached. Then, if someone asked us about the type of the converted term,
we'd go into the attachment, fetch its scala.reflect type, convert it to scala.meta
and be done with it.

Despite being very simple, this approach was also remarkable for being
an evolutionary dead-end, because storing platform-dependent information
in platform-independent trees isn't going to scale. What if we need to serialize
this information? What if we need to mix trees from different platforms?
And so on, and so forth.

So, a new approach that I put together in the day before my ScalaDays Amsterdam talk
consisted in adding a dedicated `typing` field to all ast nodes that inherit
from Term. This field would be essentially Option[Type], either being empty
or carrying a scala.meta (i.e. platform-independent) Type.

Now, instead of converting the types lazily, when they are requested,
I had to convert the types eagerly, when the corresponding trees are converted.
And that brought up a really old problem discovered by @mutcianm:
what type should we assign to method references, e.g. as in:

def foo[T](x: T) = x
foo[Int](42)

The old idea was to throw an exception if someone calls `.tpe` on `foo[Int]`
or `foo`, because those are not value types and hence can't be represented
in Scala syntax and hence shouldn't be supported by scala.meta.

This was more or less fine in the old vision of scala.meta, where it was
an unopinionated metaprogramming framework. However, in the new scheme of things
where scala.meta is optimized towards analysis of fully typechecked code
(loaded from TASTY), i.e. where all the term nodes already have their types
assigned successfully (that is, where .tpe can't fail because a tree node
is untyped or refers to an unresolved overloaded method), I think it makes
sense to try to assign types to all terms.

To do that, we need to think what to do with `foo[Int]` and `foo` in the example
above (something that corresponds to MethodType and PolyType in scalac;
there are other non-value types in scalac, e.g. ClassInfoType or OverloadedType,
but they are not relevant in our setting).

After some deliberation, I decided to go ahead and approximate MethodTypes
as corresponding function types and create a dedicated AST node for PolyTypes.
Even though there's no surface syntax for polytypes in scalac, there's
always kind-projector, and it seems that they are planning to add type lambdas
to Dotty. In any case, if necessary, we can restrict polytypes to just quasiquotes
and have the prettyprinter emit equivalent structural projects.

Let's see where this is going to lead us.